Move the SPDXLicenseKnowledgeBase to core-workflow-steps
