+++ core/core-workflow-steps/src/test/java/org/eclipse/sw360/antenna/knowledgebase/SPDXLicenseKnowledgeBaseTest.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.knowledgebase.spdx;
+package org.eclipse.sw360.antenna.knowledgebase;
 
 import org.eclipse.sw360.antenna.api.IProcessingReporter;
 import org.eclipse.sw360.antenna.model.reporting.MessageType;
